+++ frontend/src/features/display_view/chessboard.css
