+++ frontend/src/features/display_view/chessboard.tsx
